Refactor application startup and configuration
- Updated Dockerfile to expose port 8091.
- Refactored `Program.cs` to use the new WebApplication builder pattern, integrating service configuration and middleware setup.
- Removed `Startup.cs` and migrated its functionality to `Program.cs`.
- Introduced `KestrelSettings` class for server configuration.
