+++ src/AdminSite/Dockerfile
@@ -5,6 +5,7 @@ FROM mcr.microsoft.com/dotnet/aspnet:9.0 AS base
 USER $APP_UID
 WORKDIR /app
 EXPOSE 8090
+EXPOSE 8091
 
 
 # This stage is used to build the service project
